Поправлены типы данных при создании БД менеджером

+++ MetricsManager/DAL/Migrations/FirstMigration.cs
@@ -12,27 +12,27 @@ namespace MetricsManager.DAL.Migrations
                 .WithColumn("AgentUrl").AsString();
             Create.Table("CpuMetrics")
                 .WithColumn("Id").AsInt64().PrimaryKey().Identity()
-                .WithColumn("AgentId").AsInt32().ForeignKey("Agents", "AgentId")
+                .WithColumn("AgentId").AsInt64().ForeignKey("Agents", "AgentId")
                 .WithColumn("Value").AsInt32()
                 .WithColumn("Time").AsInt64();
             Create.Table("DotNetMetrics")
                 .WithColumn("Id").AsInt64().PrimaryKey().Identity()
-                .WithColumn("AgentId").AsInt32().ForeignKey("Agents", "AgentId")
+                .WithColumn("AgentId").AsInt64().ForeignKey("Agents", "AgentId")
                 .WithColumn("Value").AsInt32()
                 .WithColumn("Time").AsInt64();
             Create.Table("HddMetrics")
                 .WithColumn("Id").AsInt64().PrimaryKey().Identity()
-                .WithColumn("AgentId").AsInt32().ForeignKey("Agents", "AgentId")
+                .WithColumn("AgentId").AsInt64().ForeignKey("Agents", "AgentId")
                 .WithColumn("Value").AsInt32()
                 .WithColumn("Time").AsInt64();
             Create.Table("NetworkMetrics")
                 .WithColumn("Id").AsInt64().PrimaryKey().Identity()
-                .WithColumn("AgentId").AsInt32().ForeignKey("Agents", "AgentId")
+                .WithColumn("AgentId").AsInt64().ForeignKey("Agents", "AgentId")
                 .WithColumn("Value").AsInt32()
                 .WithColumn("Time").AsInt64();
             Create.Table("RamMetrics")
                 .WithColumn("Id").AsInt64().PrimaryKey().Identity()
-                .WithColumn("AgentId").AsInt32().ForeignKey("Agents", "AgentId")
+                .WithColumn("AgentId").AsInt64().ForeignKey("Agents", "AgentId")
                 .WithColumn("Value").AsInt32()
                 .WithColumn("Time").AsInt64();
         }
